backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.2.3
numpy==2.1.2
scikit-learn==1.3.2
joblib==1.3.2
requests==2.32.3
python-dotenv==1.0.1
matplotlib==3.9.2
seaborn==0.13.0
plotly==5.24.1
aiohttp==3.9.5
